Moved build artifacts to obj/, and restricted build/ artifacts to only the useful ones.
