+++ packages/perspective-jupyterlab/src/js/embed.js
@@ -8,7 +8,7 @@
  */
 
 import worker from "worker-loader?inline=true&fallback=false!@jpmorganchase/perspective/src/js/perspective.wasm.js";
-import buffer from "arraybuffer-loader!@jpmorganchase/perspective/build/_psp.async.wasm" 
+import buffer from "arraybuffer-loader!@jpmorganchase/perspective/build/psp.async.wasm" 
 
 window.__PSP_WORKER__ = worker;
 window.__PSP_WASM__ = buffer;
