+++ packages/perspective/CMakeLists.txt
@@ -155,21 +155,21 @@ if (PSP_WASM_BUILD)
 	add_executable(perspective.async src/cpp/main.cpp)
 	target_link_libraries(perspective.async psp "${ASYNC_MODE_FLAGS}")
 	set_target_properties(perspective.async PROPERTIES COMPILE_FLAGS "${ASYNC_MODE_FLAGS}")
-	set_target_properties(perspective.async PROPERTIES RUNTIME_OUTPUT_DIRECTORY "../build/")
-	set_target_properties(perspective.async PROPERTIES OUTPUT_NAME "_psp.async")
+	set_target_properties(perspective.async PROPERTIES RUNTIME_OUTPUT_DIRECTORY "../obj/")
+	set_target_properties(perspective.async PROPERTIES OUTPUT_NAME "psp.async")
 
 	add_executable(perspective.sync src/cpp/main.cpp)
 	target_link_libraries(perspective.sync psp "${SYNC_MODE_FLAGS}")
 	set_target_properties(perspective.sync PROPERTIES COMPILE_FLAGS "${SYNC_MODE_FLAGS}")
-	set_target_properties(perspective.sync PROPERTIES RUNTIME_OUTPUT_DIRECTORY "../build/")
-	set_target_properties(perspective.sync PROPERTIES OUTPUT_NAME "_psp.sync")
+	set_target_properties(perspective.sync PROPERTIES RUNTIME_OUTPUT_DIRECTORY "../obj/")
+	set_target_properties(perspective.sync PROPERTIES OUTPUT_NAME "psp.sync")
 	add_dependencies(perspective.sync perspective.async)
 
 	add_executable(perspective.asm src/cpp/main.cpp)
 	target_link_libraries(perspective.asm psp "${ASMJS_MODE_FLAGS}")
 	set_target_properties(perspective.asm PROPERTIES COMPILE_FLAGS "${ASMJS_MODE_FLAGS}")
-	set_target_properties(perspective.asm PROPERTIES RUNTIME_OUTPUT_DIRECTORY "../build/")
-	set_target_properties(perspective.asm PROPERTIES OUTPUT_NAME "_psp.asmjs")
+	set_target_properties(perspective.asm PROPERTIES RUNTIME_OUTPUT_DIRECTORY "../obj/")
+	set_target_properties(perspective.asm PROPERTIES OUTPUT_NAME "psp.asmjs")
 	add_dependencies(perspective.asm perspective.sync)
 endif()
