+++ packages/perspective/package.json
@@ -19,9 +19,11 @@
     "bench:build:copy": "cp bench/csv/* build",    
     "bench:run": "node build/benchmark.js",
     "bench": "npm-run-all bench:build bench:run",
-    "build": "npm-run-all build:compile:* build:webpack",
+    "build": "npm-run-all build:compile:* build:copy:* build:webpack",
     "build:compile:copy": "mkdir -p obj build",
     "build:compile:emmake": "cd obj/ && emcmake cmake ../ && emmake make -j${PSP_CPU_COUNT-8}",
+    "build:copy:async": "cp obj/psp.async.wasm build/psp.async.wasm",
+    "build:copy:sync": "cp obj/psp.sync.wasm build/psp.sync.wasm",
     "build:webpack": "npm-run-all -p build:webpack:*",
     "build:webpack:asmjs": "webpack --color --config src/config/perspective.asmjs.config.js",
     "build:webpack:wasm": "webpack --color --config src/config/perspective.wasm.config.js",
